Fix release notes view when there are multiple files attached to the version
